// src/Models/ResponseModels/GetAllUserDetails.ts
export class GetAllUserDetails {
  id: string = "";
  firstName: string = "";
  surname: string = "";
  displayName: string = "";
  email: string = "";
  username: string = "";
  phoneNumber: number = 0;
  createdAt: string = "";
  statusId: string = "";
  isActive: boolean = false;
  roleName: string = "";
  status: string = "";
}